changed position type from int to uint8_t

+++ enclave/lib/App/ecu.h
@@ -27,18 +27,20 @@ typedef struct _ecu_t {
   cp_ec256_private_t ecu_priv_key;
 } ecu_t;
 
-commpact_status_t setEnclavePubKey(int position, cp_ec256_public_t *pub_key);
-commpact_status_t setEnclavePubKeyRealECU(int position,
+commpact_status_t setEnclavePubKey(uint8_t position,
+                                   cp_ec256_public_t *pub_key);
+commpact_status_t setEnclavePubKeyRealECU(uint8_t position,
                                           cp_ec256_public_t *pub_key);
-commpact_status_t getRealECUPubKey(int position, cp_ec256_public_t *pub_key);
-commpact_status_t setParametersECU(int position,
+commpact_status_t getRealECUPubKey(uint8_t position,
+                                   cp_ec256_public_t *pub_key);
+commpact_status_t setParametersECU(uint8_t position,
                                    cp_ec256_signature_t *enclave_signature,
                                    ecu_message_t *message,
                                    cp_ec256_signature_t *ecu_signature);
-commpact_status_t generateKeyPair(int position, cp_ec256_public_t *pub_key);
+commpact_status_t generateKeyPair(uint8_t position, cp_ec256_public_t *pub_key);
 commpact_status_t signMessage(ecu_t *ecu, ecu_message_t *message,
                               cp_ec256_signature_t *signature);
-commpact_status_t setParametersRealECU(int position,
+commpact_status_t setParametersRealECU(uint8_t position,
                                        cp_ec256_signature_t *enclave_signature,
                                        ecu_message_t *message,
                                        cp_ec256_signature_t *ecu_signature);
